[#37]Refactor(comment): 폴더명 변경
