+++ src/main/java/today/sesac/shoutify/comment/repository/CommentRepository.java
@@ -2,7 +2,8 @@ package today.sesac.shoutify.comment.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import today.sesac.shoutify.comment.domain.Comment;
+
+import today.sesac.shoutify.comment.entity.Comment;
 
 @Repository
 public interface CommentRepository extends JpaRepository<Comment, Long> {
